% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{allequalService}
\alias{allequalService}
\title{Grade student Rmds based on all.equal messages}
\usage{
allequalService(
  targetLabel = targetLabel,
  .transform = NULL,
  switchTargetCurrent = F,
  useSHA = F
)
}
\value{

}
\description{
Service initiator for all.equal comparisons
}
\examples{
none
}
