% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal2.R
\name{allequalService2}
\alias{allequalService2}
\title{Initiate allequal service 2}
\usage{
allequalService2(process, path = "")
}
\arguments{
\item{process}{a Process instance}

\item{path}{The path where student Rmds sit.}
}
\value{

}
\description{
This is an enhanced version of allequalService. The instance is equipped with file.edit method to bring out student Rmd, and print_code method to print any particular code of any student's Rmd
}
\examples{
none
}
