% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{augment_studentRmdsWithAtAnsFunctional}
\alias{augment_studentRmdsWithAtAnsFunctional}
\title{Generate the funtion that can combine correct answer and student answer into one Rmd}
\usage{
augment_studentRmdsWithAtAnsFunctional(ansRmdAccommodateAtAns, ansChunkTable)
}
\arguments{
\item{ansRmdAccommodateAtAns}{A character of lines from ansRmd that already accommodate \%at\% ans}

\item{ansChunkTable}{A data frame that show code chunk mapping of ansRmdAccommodateAtAns}
}
\value{

}
\description{
Generate the funtion that can combine correct answer and student answer into one Rmd
}
\examples{
none
}
