% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{browseStudentCourseworkWebpageService}
\alias{browseStudentCourseworkWebpageService}
\title{Build student coursework webpage browsing service.}
\usage{
browseStudentCourseworkWebpageService()
}
\value{

}
\description{
The function create a courseWorkService environment in global and bring out finder for you to choose files from dispute folder. Once copy file name. Run  courseWorkService$build_courseworkBrowse() to complete the service build
}
\examples{
none
}
