% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{categorise_elementNames_ByElementValues}
\alias{categorise_elementNames_ByElementValues}
\title{Categorise a character vector's element names by its element values}
\usage{
categorise_elementNames_ByElementValues(rmdNamed_char)
}
\arguments{
\item{rmdNamed_char.}{A named vector.}
}
\value{

}
\description{
Categorise a character vector's element names by its element values
}
\examples{

x <- sample(1:5, 10, replace = T)
names(x) <- sample(LETTERS, 10)
cat_x <-
  categorise_elementNames_ByElementValues(x)
}
