% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{categorize_targetIs}
\alias{categorize_targetIs}
\title{Divide list of all.equal msg into list of group with different names of "Target is..."}
\usage{
categorize_targetIs(summary31b)
}
\arguments{
\item{summary31b}{A list of all.equal msgs}
}
\value{

}
\description{
Divide list of all.equal msg into list of group with different names of "Target is..."
}
\examples{
none
}
