% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{compute_gradesFromMultipleAes}
\alias{compute_gradesFromMultipleAes}
\title{When facing list of ae's, use this to compute grades as sum/mean across all ae's extracted grades}
\usage{
compute_gradesFromMultipleAes(list_ae)
}
\arguments{
\item{list_ae}{A list of ae's.}
}
\value{
A data frame with sum and mean of grades across all versions of ae's
}
\description{
When facing list of ae's, use this to compute grades as sum/mean across all ae's extracted grades
}
\examples{
none
}
