% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeComputation.R
\name{convert_gradeList2dataframe}
\alias{convert_gradeList2dataframe}
\title{Convert list of grades into a data frame with total computed}
\usage{
convert_gradeList2dataframe(
  ...,
  turnInBonus = 3,
  maxPoint = 10,
  fullMark = NULL
)
}
\arguments{
\item{...}{A name=value pair, such as ans11=grade11, ans12=grade12,...}
}
\value{

}
\description{
Convert list of grades into a data frame with total computed
}
\examples{
none
}
