% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{copy_toGoogleDriveReturnFolder}
\alias{copy_toGoogleDriveReturnFolder}
\title{Copy list of return Rmds to their individual google drivel return folders}
\usage{
copy_toGoogleDriveReturnFolder(
  listRmdsInReturnFolder,
  title,
  localGDReturnFolderPath
)
}
\arguments{
\item{listRmdsInReturnFolder}{A character of the returnRmds with full path}

\item{title}{A character usually the name of the exercise such as hw1, midterm1, etc...}

\item{localGDReturnFolderPath}{A character of path to local google drive return folder}
}
\value{

}
\description{
Copy list of return Rmds to their individual google drivel return folders
}
\examples{
none
}
