% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{copy_toGoogleDriveReturnFolder2}
\alias{copy_toGoogleDriveReturnFolder2}
\title{Copy list of return Rmds to their individual google drivel return folders}
\usage{
copy_toGoogleDriveReturnFolder2(listRmdsInReturnFolder, title)
}
\arguments{
\item{listRmdsInReturnFolder}{A character of the returnRmds with full path}

\item{title}{A character usually the name of the exercise such as hw1, midterm1, etc...}
}
\value{

}
\description{
Copy list of return Rmds to their individual google drivel return folders
}
\examples{
none
}
