% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{copyCorrectAnsRmdToGDCannotSynthesizedFolder}
\alias{copyCorrectAnsRmdToGDCannotSynthesizedFolder}
\title{Generate ansRmd with \%at\% ans and copy to individual return folder in google drive}
\usage{
copyCorrectAnsRmdToGDCannotSynthesizedFolder(
  correctAnsTemplateFilename,
  listRmdsInCannotSynthesizedFolder,
  title,
  localGDReturnFolderPath,
  turnInBonus
)
}
\arguments{
\item{correctAnsTemplateFilename}{A character of the full path to the correct answer template Rmd}

\item{listRmdsInCannotSynthesizedFolder}{A character of the full path to the folder storing all cannotSynthesized Rmds}

\item{title}{A character such as hw1, midterm1, etc}

\item{localGDReturnFolderPath}{A character of path to local google drive return folder}
}
\value{

}
\description{
Generate ansRmd with \%at\% ans and copy to individual return folder in google drive
}
\examples{
none
}
