% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{copyCorrectAnsRmdToGDCannotSynthesizedFolder2}
\alias{copyCorrectAnsRmdToGDCannotSynthesizedFolder2}
\title{Generate ansRmd with \%at\% ans and copy to individual return folder in google drive}
\usage{
copyCorrectAnsRmdToGDCannotSynthesizedFolder2(
  correctAnsTemplateFilename,
  listRmdsInCannotSynthesizedFolder,
  title
)
}
\arguments{
\item{correctAnsTemplateFilename}{A character of the full path to the correct answer template Rmd}

\item{listRmdsInCannotSynthesizedFolder}{A character of the full path to the folder storing all cannotSynthesized Rmds}

\item{title}{A character such as hw1, midterm1, etc}
}
\value{

}
\description{
Generate ansRmd with \%at\% ans and copy to individual return folder in google drive
}
\examples{
none
}
