% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{create_individualGoogleDriveReturnFolder}
\alias{create_individualGoogleDriveReturnFolder}
\title{Create individual return folders under class google drive folder}
\usage{
create_individualGoogleDriveReturnFolder(classDriveFolderPath, roster)
}
\arguments{
\item{classDriveFolderPath}{A character of the url to the class google drive folder}

\item{roster}{A data frame with 學號, used to create individual return gd folder}
}
\value{
A list of created gd folder details
}
\description{
Create individual return folders under class google drive folder
}
\examples{
none
}
