% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{decomposeObject}
\alias{decomposeObject}
\title{Decompose an object for systematic grading}
\usage{
decomposeObject(y)
}
\arguments{
\item{y}{an object}
}
\value{
a list of class(object), length(object), el_withNames, el_noName
}
\description{
Decompose an object for systematic grading
}
\examples{
none
}
