% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{eval_safe}
\alias{eval_safe}
\title{evaluate language (any code within \{\}) safely}
\usage{
eval_safe(language, errorReturn = NA, envir = rlang::caller_env())
}
\arguments{
\item{language}{A language. Any codes as long as inside \{\}}

\item{errorReturn}{default=NA. What would be returned when the language can not be evaluated.}

\item{envir}{default=caller environment. Where to evaluate the language}
}
\value{

}
\description{
evaluate language (any code within \{\}) safely
}
\examples{
eval_safe({a <- 3; b<-a+4})
}
