% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation2.R
\name{Evaluate2}
\alias{Evaluate2}
\title{Initiate an Evaluation instance after Process instance is done}
\usage{
Evaluate2(pe)
}
\arguments{
\item{pe}{A Process instance that has been processed.}
}
\value{

}
\description{
Initiate an Evaluation instance after Process instance is done
}
\examples{
\dontrun{
pe <- Process()
ei <- Evalutate(pe)
}
}
