% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{execute_mgetxy}
\alias{execute_mgetxy}
\title{Execute mgetxy for some ansXXX with transform function possibility}
\usage{
execute_mgetxy(mgetxy, targetLabel, .transform = NULL)
}
\arguments{
\item{mgetxy}{An mgetxy object}

\item{targetLabel}{A ansXXX label}

\item{.transform}{default=NULL, a transform function}
}
\value{
a list of xy
}
\description{
Execute mgetxy for some ansXXX with transform function possibility
}
\examples{
none
}
