% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{extract_exerciseRmd_from_ansRmd}
\alias{extract_exerciseRmd_from_ansRmd}
\title{With XXX-ans.Rmd file open, this extract exercise Rmd for student to do with no answer}
\usage{
extract_exerciseRmd_from_ansRmd()
}
\value{

}
\description{
With XXX-ans.Rmd file open, this extract exercise Rmd for student to do with no answer
}
\examples{
none
}
