% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{filterMessage_Functional}
\alias{filterMessage_Functional}
\title{Subset list of all.equal messages down to those fit msgKeyphrase pattern}
\usage{
filterMessage_Functional(list_allequalMsgs)
}
\arguments{
\item{list_allequalMsgs}{A list of all.equal msg to be refined}
}
\value{
A refined list
}
\description{
Subset list of all.equal messages down to those fit msgKeyphrase pattern
}
\examples{
none
}
