% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{generate_disputeFilesFromIssueComments}
\alias{generate_disputeFilesFromIssueComments}
\title{Generate dispute files from issue comments}
\usage{
generate_disputeFilesFromIssueComments(
  issueComments,
  title,
  sourceReturnFolder,
  disputeFolder,
  overwrite = c(dupReturnRmd = F, disputeContentRmd = F)
)
}
\arguments{
\item{issueComments}{A list of issue comments returned by obtainIssueComments()}

\item{title}{A character of exercise title, such as hw1, midterm1}

\item{sourceReturnFolder}{A character of the path where the returned Rmds are stored}

\item{disputeFolder}{A character of the path where the generated dispute files are to be placed at}

\item{overwrite}{A logical vector of 2, first element T= overwrite duplicated return Rmd, second element T= overwrite dispute content Rmd#'}

\item{commentTarget}{A list of comment return from thread$comments_url$get_comments()}
}
\value{

}
\description{
Generate dispute files from issue comments
}
\examples{
none
}
