% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{generate_disputeFilesFromOneComment}
\alias{generate_disputeFilesFromOneComment}
\title{Generate marked Rmd and its dispute Rmd for dispute resolution}
\usage{
generate_disputeFilesFromOneComment(
  commentTarget,
  title,
  sourceReturnFolder,
  disputeFolder,
  overwrite = c(F, F)
)
}
\arguments{
\item{commentTarget}{A list of comment return from thread$comments_url$get_comments()}

\item{title}{A character of exercise title, such as hw1, midterm1}

\item{sourceReturnFolder}{A character of the path where the returned Rmds are stored}

\item{disputeFolder}{A character of the path where the generated dispute files are to be placed at}

\item{overwrite}{A logical vector of 2, first element T= overwrite duplicated return Rmd, second element T= overwrite dispute content Rmd}
}
\value{

}
\description{
Generate marked Rmd and its dispute Rmd for dispute resolution
}
\examples{
none
}
