% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{generate_mgetxy}
\alias{generate_mgetxy}
\title{Generate a list of getxy from a group of Rmds selected}
\usage{
generate_mgetxy(eliteGroup)
}
\arguments{
\item{eliteGroup}{A character vector each represent a chosen Rmd to form getxy as how we use in getxy input argument. This is a vectorized version.}
}
\value{

}
\description{
Generate a list of getxy from a group of Rmds selected
}
\examples{
none
}
