% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{generate_recordGradeCommentsFromTb_gradesWithoutComments}
\alias{generate_recordGradeCommentsFromTb_gradesWithoutComments}
\title{Generate records_gradeComment for step return from tb_grades}
\usage{
generate_recordGradeCommentsFromTb_gradesWithoutComments(tb_grades)
}
\arguments{
\item{tb_grades}{A data frame.}
}
\value{

}
\description{
Generate records_gradeComment for step return from tb_grades
}
\examples{
none
}
