% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatingChart.R
\name{generate_seatingChartWebpage}
\alias{generate_seatingChartWebpage}
\title{Generate a webpage for student to inquire the assigned seat}
\usage{
generate_seatingChartWebpage(
  seatingChartUrl,
  sheet_seatChart,
  range_seatChart,
  tag_iframe,
  file
)
}
\arguments{
\item{seatingChartUrl}{A url address to google sheet}

\item{sheet_seatChart}{A sheet name of the seating chart}

\item{range_seatChart}{A range expression of seating chart}

\item{tag_iframe}{A html tag of iframe generated by google sheet embed publishing}

\item{file}{A filepath to which the rendered html file will be saved}
}
\value{

}
\description{
Generate a webpage for student to inquire the assigned seat
}
\examples{
none
}
