% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNonAnsRmd.R
\name{generate_studentRmdFromAnsRmd}
\alias{generate_studentRmdFromAnsRmd}
\title{Generate homework or exam Rmd for students from ans Rmd}
\usage{
generate_studentRmdFromAnsRmd(ansFilename, destfolderPath = ".")
}
\arguments{
\item{ansFilename}{A character of full path to the ans Rmd}
}
\value{
A list of processed chunks. When unlist, it will be the student Rmd content.
}
\description{
ansXX label code chunks will be processed. Block programming will preserve the outer most block. The last line will always be the markoff answer object based on the object shown in the last line in ans Rmd
}
\examples{
none
}
