% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{generateReturnRmds_then_copy2googleDriveReturnFolder}
\alias{generateReturnRmds_then_copy2googleDriveReturnFolder}
\title{Generate student-correct answer synthesized Rmd and copy to each student's return folder}
\usage{
generateReturnRmds_then_copy2googleDriveReturnFolder(
  title,
  roster,
  correctAnsTemplateFilename,
  localGDReturnFolderPath,
  turnInBonus = 3
)
}
\arguments{
\item{title}{A character of exercise title such as hw1, midterm1, etc}

\item{roster}{A data frame with 學號, Gmail帳號}

\item{correctAnsTemplateFilename}{A character of filepath to the correct answer Rmd template to used be used for synthesizing}

\item{localGDReturnFolderPath}{A character of path to local google drive return folder}
}
\value{

}
\description{
Generate student-correct answer synthesized Rmd and copy to each student's return folder
}
\examples{
none
}
