% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{get_allequalSummary}
\alias{get_allequalSummary}
\title{Generate list of all.equal results where Target is correct answer and Current is student answer}
\usage{
get_allequalSummary(
  targetLabel,
  whichCorrectAnsvalue = 1,
  .transform = NULL,
  switchTargetCurrent = F
)
}
\arguments{
\item{targetLabel}{A character of like ans11, etc}

\item{whichCorrectAnsvalue}{A number default at 1, indicating which values in correct answer values is used for comparison}

\item{.transform}{A function, default at NULL. If not null, it is the function that transforms x and y before grading and comparison}
}
\value{
A list of all.equal outcomes
}
\description{
Generate list of all.equal results where Target is correct answer and Current is student answer
}
\examples{
none
}
