% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_answerObjectValues}
\alias{get_answerObjectValues}
\title{Execute Rmd scripts and obtain answer object values}
\usage{
get_answerObjectValues(studentsRmds, correctAnsFilename)
}
\arguments{
\item{studentsRmds}{A character of full filename path of Rmds}

\item{correctAnsFilename}{A character of full filename path of correct Answer Rmds}
}
\value{
A list of 2
}
\description{
Execute Rmd scripts and obtain answer object values
}
\examples{
none
}
