% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatingChart.R
\name{get_seatmapFromSeatChart}
\alias{get_seatmapFromSeatChart}
\title{Recover seat mapping from a seating chart google sheet}
\usage{
get_seatmapFromSeatChart(seatingChartUrl, sheet_seatChart, range_seatChart)
}
\arguments{
\item{seatingChartUrl}{A url address to google sheet}

\item{sheet_seatChart}{A sheet name of the seating chart}

\item{range_seatChart}{A range expression of seating chart}
}
\value{

}
\description{
Recover seat mapping from a seating chart google sheet
}
\examples{
none
}
