% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{get_vectorOfAllPossibleMsgs}
\alias{get_vectorOfAllPossibleMsgs}
\title{Turn list of all.equal msg to unique msg vectors}
\usage{
get_vectorOfAllPossibleMsgs(summary31b)
}
\arguments{
\item{summary31b}{list of all.equal msgs}
}
\value{

}
\description{
Turn list of all.equal msg to unique msg vectors
}
\examples{
none
}
