% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeComputation.R
\name{getMaxGrade}
\alias{getMaxGrade}
\title{get max grade out of grade vectors}
\usage{
getMaxGrade(...)
}
\arguments{
\item{...}{grade vectors with rmd as element names}
}
\value{
a grade vector with rmd as element names
}
\description{
get max grade out of grade vectors
}
\examples{
\dontrun{
getMaxGrade(grade111, grade112, grade113)
}
}
