% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{getxyFunctional}
\alias{getxyFunctional}
\title{Generate function to quickly pull out specific answers from a student and the correcnt answers}
\usage{
getxyFunctional(.x)
}
\arguments{
\item{.x}{A character or number. When character, it should be from names(studentValues); When number, it should be the location regarding element in studentValues.}
}
\value{
A function
}
\description{
Given student index in studentValues, generate a getxy function of answer label which quickly gives x(this student's answer) and y(correct answer) in the global environment
}
\examples{
none.
}
