% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{grade_by_all.equalMessages}
\alias{grade_by_all.equalMessages}
\title{Grade student Rmds based on all.equal messages}
\usage{
grade_by_all.equalMessages(ae)
}
\arguments{
\item{ae}{An all.equal_env object after assigning grades for all messages groups}
}
\value{
A list of grades and comments
}
\description{
Grade student Rmds based on all.equal messages
}
\examples{
none
}
