% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_factor.R
\name{grade_factorFunctional}
\alias{grade_factorFunctional}
\title{Generate a grading function for factor vector}
\usage{
grade_factorFunctional(point_factor)
}
\arguments{
\item{point_factor}{A list follows the template return from point_factorTemplate()}
}
\value{
A function whose implementation will yield a list of point and details
}
\description{
Generate a grading function for factor vector
}
\examples{
none.
}
