% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grademFunctional}
\alias{grademFunctional}
\title{Grade function specific to elite group}
\usage{
grademFunctional(.mgetxy)
}
\arguments{
\item{.mgetxy}{A list of getxy functions}
}
\value{

}
\description{
Grade function specific to elite group
}
\examples{
none
}
