% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{gradeMix}
\alias{gradeMix}
\title{A mixture of m grading and all.equal grading}
\usage{
gradeMix(ae, gradingMethod, whichCorrectAnsvalue = 1)
}
\arguments{
\item{ae}{A all.equal_env object, which has .yield_messageGroupTable method already applied.}

\item{gradingMethod}{A function of m grading}

\item{whichCorrectAnsvalue}{An integer indicating which object value to be used as answer}
}
\value{
a list of grades with Rmd names as element names
}
\description{
A mixture of m grading and all.equal grading
}
\examples{
none
}
