% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{Group}
\alias{Group}
\title{Instance initiator to convenience grouping task}
\usage{
Group(vector2group)
}
\arguments{
\item{vector2group}{A vector of characters to be grouped.}
}
\value{
An environment instance for grouping
}
\description{
Instance initiator to convenience grouping task
}
\examples{
none
}
