% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{ifelsethen}
\alias{ifelsethen}
\title{The extension to ifelse to accommodate error situation}
\usage{
ifelsethen(cond, v1, v2, errorMsg)
}
\arguments{
\item{cond}{A logical}

\item{v1}{value to return when T}

\item{v2}{value to return when F}

\item{errorMsg}{message to print and return when error}
}
\value{

}
\description{
The extension to ifelse to accommodate error situation
}
\examples{
none
}
