% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{Inspect}
\alias{Inspect}
\title{Create inspection instance when df_grades in the global environment}
\usage{
Inspect(path)
}
\arguments{
\item{path}{A path to the folder where all the synthesized Rmds are.}
}
\value{

}
\description{
Create inspection instance when df_grades in the global environment
}
\examples{
none.
}
