% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{inspect_selectedDisputeFile}
\alias{inspect_selectedDisputeFile}
\title{Select disput files from menu and investigate}
\usage{
inspect_selectedDisputeFile(disputeFolder, disputeFileInfo)
}
\arguments{
\item{disputeFolder}{A character of dispute folder path}

\item{disputeFileInfo}{A list of return from generate_disputeFilesFromIssueComments}
}
\value{

}
\description{
Select disput files from menu and investigate
}
\examples{
none.
}
