% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeComputation.R
\name{merge_gradeVectors}
\alias{merge_gradeVectors}
\title{Merge several named numerical \\(grade\\) vectors into a tibble}
\usage{
merge_gradeVectors(...)
}
\arguments{
\item{...}{arguments of grade vectors}
}
\value{
A tibble with Rmd column as the merging key.
}
\description{
Merge several named numerical \\(grade\\) vectors into a tibble
}
\examples{
\dontrun{
merge_gradeVectors(grade111, grade112, grade113)
}
}
