% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_answerEnvironment.R
\name{mget_safe}
\alias{mget_safe}
\title{a safe version of mget, this allow input argument to be either a character of object name or a character of obtaining value, such as "a[['number']][[1]]"}
\usage{
mget_safe(text, envir = rlang::caller_env())
}
\arguments{
\item{text}{A character either represents an object name or how to obtain a value}

\item{envir}{An environment where to look for value}
}
\value{

}
\description{
a safe version of mget, this allow input argument to be either a character of object name or a character of obtaining value, such as "a[['number']][[1]]"
}
\examples{
none
}
