% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{obtainIssueComments}
\alias{obtainIssueComments}
\title{Obtain issue comments}
\usage{
obtainIssueComments(owner, repo, titleKeyword, .label = "dispute")
}
\arguments{
\item{owner}{A character of github username}

\item{repo}{A character of repo name}

\item{titleKeyword}{A character of issue title keyword}

\item{.label}{A character of label of issue to look for, default="grievance".}
}
\value{
A list of comments
}
\description{
Obtain issue comments
}
\examples{
none
}
