% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_factor.R
\name{point_factorTemplate}
\alias{point_factorTemplate}
\title{Generate a list of point distribution template for using in grade_factorFunctional()}
\usage{
point_factorTemplate()
}
\value{
A list of point distribution the point element values should summed up to one. As to weight, it is the weight applied to the point in each situation.
}
\description{
Generate a list of point distribution template for using in grade_factorFunctional()
}
\examples{
none
}
