% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{post_disputeIssue2Github}
\alias{post_disputeIssue2Github}
\title{Post the dispute issue}
\usage{
post_disputeIssue2Github(repoUrl, title, content, label = "dispute")
}
\arguments{
\item{repoUrl}{A character of repo url}

\item{title}{A character}

\item{content}{A character}

\item{label}{A character default="dispute"}
}
\value{

}
\description{
Post the dispute issue
}
\examples{
none
}
