% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{Process}
\alias{Process}
\title{Initiator for Rmd processing service}
\usage{
Process()
}
\value{
an processing instance environment
}
\description{
Initiator for Rmd processing service
}
\examples{
\dontrun{
process <- Process()
}
}
