% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process2.R
\name{Process2}
\alias{Process2}
\title{Initiator for Rmd processing service}
\usage{
Process2(ansRmd, path_studentRmds)
}
\value{
an processing instance environment
}
\description{
Initiator for Rmd processing service
}
\examples{
\dontrun{
process <- Process2()
}
}
