% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeJSON.R
\name{read_googleDriveJson}
\alias{read_googleDriveJson}
\title{Read a json file from Google Drive url}
\usage{
read_googleDriveJson(fileUrl)
}
\arguments{
\item{fileUrl}{A weblink to a google drive jsonfile}
}
\value{

}
\description{
Read a json file from Google Drive url
}
\examples{
none
read_googleDrivJson("https://drive.google.com/file/d/17D0TYnwJKps1ZxrbHvENs9LcSOY1JtXA/view?usp=sharing")
}
