% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{record_gradesCommentsWithTimestamp}
\alias{record_gradesCommentsWithTimestamp}
\title{Record student grades and comments as a list of many students}
\usage{
record_gradesCommentsWithTimestamp(all_aeObjects, envir = rlang::current_env())
}
\arguments{
\item{all_aeObjects}{A character vector of ae for each ansXX}

\item{envir}{An environment where all_aeObjects names exist}
}
\value{
a list of many students, each is a list of many ansXX which has 3 elements, timestamp, grade, comment
}
\description{
Record student grades and comments as a list of many students
}
\examples{
none
}
