% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{recordAnsObjValues_withFlagExecutable}
\alias{recordAnsObjValues_withFlagExecutable}
\title{With executable flag, record the evaluated ans object values in answerEnvironment}
\usage{
recordAnsObjValues_withFlagExecutable(
  flag_executable,
  targetAnsLabel,
  saveEnv,
  evalEnv = NULL
)
}
\arguments{
\item{flag_executable}{A logical}

\item{targetAnsLabel}{A character}

\item{saveEnv}{An environment to save ans object value}

\item{evalEnv}{An environment where to evaluate expressions}
}
\value{
none, but saveEnv is updated
}
\description{
With executable flag, record the evaluated ans object values in answerEnvironment
}
\examples{
none
}
