% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{Return}
\alias{Return}
\title{Return instance environment generator}
\usage{
Return(pe, returnFolderpath)
}
\arguments{
\item{pe}{An process instance from Process instance initiator}

\item{returnFolderpath}{A path under which all students' return folders lie in}
}
\value{

}
\description{
Initiate a return process instance.
}
\examples{
none
}
