% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return2.R
\name{Return2}
\alias{Return2}
\title{Generate return instance to return Rmd in Google classroom}
\usage{
Return2(path, download)
}
\arguments{
\item{path}{a path to the folder where all the Rmds there are to be returned.}

\item{download}{an instance of Download.}
}
\value{

}
\description{
Generate return instance to return Rmd in Google classroom
}
\examples{
none
}
