% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{returnAllDisputeFiles}
\alias{returnAllDisputeFiles}
\title{Return all processed dispute files to students' google drive return folder}
\usage{
returnAllDisputeFiles(
  disputeFolder,
  localGDReturnFolderPath,
  title,
  overwrite = T
)
}
\arguments{
\item{disputeFolder}{A character of path to local dispute folder}

\item{localGDReturnFolderPath}{A character of path to local google drive return folder}

\item{title}{A character of exercise title}

\item{overwrite}{defaul=T}
}
\value{

}
\description{
Return all processed dispute files to students' google drive return folder
}
\examples{
none
}
