% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{returnCopiedDisputeFilesFromMacFinder}
\alias{returnCopiedDisputeFilesFromMacFinder}
\title{Return copied full path filenames to correct individual return folders}
\usage{
returnCopiedDisputeFilesFromMacFinder()
}
\value{

}
\description{
Return copied full path filenames to correct individual return folders
}
\examples{
none.
}
