% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeJSON.R
\name{Review}
\alias{Review}
\title{Review instance initiator}
\usage{
Review(tb_grades, title)
}
\arguments{
\item{tb_grades}{A dataframe of grades from each ansXXX. If need to be recale, recale them before input here.}

\item{title}{A character of title, normally from params$title}
}
\value{
A review instance
}
\description{
Review instance initiator
}
\examples{
None
}
