% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{reviseRmd_atAns}
\alias{reviseRmd_atAns}
\title{Revise Rmd to accommodate \%at\% ans}
\usage{
reviseRmd_atAns(ansRmdlines)
}
\arguments{
\item{ansRmdlines}{A character of Rmd lines}
}
\value{

}
\description{
Revise Rmd to accommodate \%at\% ans
}
\examples{
none
}
