% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmd_chunkTable}
\alias{rmd_chunkTable}
\title{Get chunk table from rmdlines}
\usage{
rmd_chunkTable(rmdlines)
}
\arguments{
\item{rmdlines}{a character vector, possibily read from xfun::read_utf8 from an Rmd file.}
}
\value{
a chunk table data frame
}
\description{
With character vectors read from an Rmd file, this function returns a data frame chunk table which tells you the start, end and chunk labels in the rmdlines
}
\examples{
\dontrun{
filename <- "/Users/martinl/Github/course-dashboard-programming-for-data-science//grading_flow/HW8/homework8-ans.Rmd"
xfun::read_utf8(process$correctAnsFilename$filename) -> rmdlines
rmd_chunkTable(rmdlines)
}
}
