% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmd_list_codeChunks}
\alias{rmd_list_codeChunks}
\title{Produce list of code chunks}
\usage{
rmd_list_codeChunks(rmdlines)
}
\arguments{
\item{rmdlines}{A character vector of rmdlines}
}
\value{
A list with elements named by chunk labels and element values of rmdlines character vector segments
}
\description{
Produce list of code chunks
}
\examples{
none.
}
