% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rootadjustPath}
\alias{rootadjustPath}
\title{Ensure the path is absolute.}
\usage{
rootadjustPath(path)
}
\arguments{
\item{path}{A character of path, absolute or relative.}
}
\value{
A character of absolute path.
}
\description{
Ensure the path is absolute.
}
\examples{
none
}
