% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{safe_unlist}
\alias{safe_unlist}
\title{Safely unlist a list where element with zero length will be list(0)}
\usage{
safe_unlist(listObj)
}
\arguments{
\item{listObj}{A list}
}
\value{

}
\description{
Safely unlist a list where element with zero length will be list(0)
}
\examples{
none
}
