% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{share_gdReturnFolders}
\alias{share_gdReturnFolders}
\title{Share individual google return folder to each student}
\usage{
share_gdReturnFolders(gd_returnFolderIndividuals)
}
\arguments{
\item{gd_returnFolderIndividuals}{A list of return from create_individualGoogleDriveReturnFolder}
}
\value{

}
\description{
Share individual google return folder to each student
}
\examples{
none
}
