% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_isFunction}
\alias{str_isFunction}
\title{Verify if a string name is a function name}
\usage{
str_isFunction(FunctionUniqueList)
}
\arguments{
\item{FunctionUniqueList}{A vector of characters}
}
\value{

}
\description{
Verify if a string name is a function name
}
\examples{
str_isFunction(c("names","name"))
}
