% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{subcategorise_byMsgRmdPairs}
\alias{subcategorise_byMsgRmdPairs}
\title{Categorise rmd based on manually defined list of msg and rmd elements}
\usage{
subcategorise_byMsgRmdPairs(ae, list_msg_rmd_pairs)
}
\arguments{
\item{ae}{An ae environment to attach subcat element}

\item{list_msg_rmd_pairs}{A list of msg and rmd elements}
}
\value{
none. but attach subcat in ae
}
\description{
Categorise rmd based on manually defined list of msg and rmd elements
}
\examples{
none
}
