% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{synthersizeWithCorrectAnsFunctional}
\alias{synthersizeWithCorrectAnsFunctional}
\title{Generate synthesizer that sythersizes both correct answer and student answer}
\usage{
synthersizeWithCorrectAnsFunctional(correctAnsFilename)
}
\arguments{
\item{correctAnsFilename}{A character of filepath to the correct answer Rmd}
}
\value{

}
\description{
the synthesizer takes in only one input that is the codeChunkTable produced by rmd2drake::get_listCodeChunksFromRmdlines
}
\examples{
none
}
