% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{Synthesize}
\alias{Synthesize}
\title{Genrate Synthesize instance to put student grades in their Rmds}
\usage{
Synthesize(df_grades, mapping, path)
}
\arguments{
\item{df_grades}{a data frame from the RDS file read generated by AllequalService2 instance's export_grade method.}

\item{mapping}{a list of ans label and Rmd section title (name-value) pair. For example, list(ans11="## 1.1", ans12="## 1.2", ans13="## 1.3", ans14="## 1.4", ans15="## 1.5")}

\item{path}{a path where the current student Rmds sit in.}
}
\value{

}
\description{
Genrate Synthesize instance to put student grades in their Rmds
}
\examples{
none
}
