% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRmd.R
\name{synthesize_returnRmd}
\alias{synthesize_returnRmd}
\title{Synthersize student Rmd and answer Rmd template to form return Rmd}
\usage{
synthesize_returnRmd(
  studentRmdfilename,
  gradeRecord,
  correctAnswerSynthesizer,
  returnFolder,
  needPR = F
)
}
\arguments{
\item{studentRmdfilename}{A character of student Rmd filepath}

\item{gradeRecord}{A data frame from the row regarding the student}

\item{correctAnswerSynthesizer}{A function generated from synthersizeWithCorrectAnsFunctional}

\item{returnFolder}{A path for return Rmd storage}

\item{needPR}{default=F. Want return Rmd to include his/her ranking.}
}
\value{

}
\description{
Synthersize student Rmd and answer Rmd template to form return Rmd
}
\examples{
none
}
