% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{transformFunctional}
\alias{transformFunctional}
\title{Generate transform function for function grading}
\usage{
transformFunctional(...)
}
\arguments{
\item{...}{A name-value pairs for function-to-grade's inputs}
}
\value{
A transform function of one functional class input
}
\description{
Generate transform function for function grading
}
\examples{
none
}
