% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{transformFunctional_messageKeep}
\alias{transformFunctional_messageKeep}
\title{Transform functional that modifies x function to keep all screen messages and turn x function into x(...)}
\usage{
transformFunctional_messageKeep(...)
}
\arguments{
\item{...}{a list-pair input arguments of original x function}
}
\value{
A list of return (returned value from x(...) call) and messages on screen
}
\description{
Transform functional that modifies x function to keep all screen messages and turn x function into x(...)
}
\examples{
none.
}
