% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{turnMsg2Null_byPattern}
\alias{turnMsg2Null_byPattern}
\title{When all.equal msg shows pattern, turn msg to NULL (i.e. error)}
\usage{
turnMsg2Null_byPattern(summary31, pattern)
}
\arguments{
\item{summary31}{list of all.equal messages}

\item{pattern}{Regex pattern}
}
\value{

}
\description{
When all.equal msg shows pattern, turn msg to NULL (i.e. error)
}
\examples{
none
}
