% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispute.R
\name{update_tbGrades}
\alias{update_tbGrades}
\title{Update tb_Grades according to disputeFolder Rmd result}
\usage{
update_tbGrades(tb_Grades, disputeFolder, disputeFileInfo, title)
}
\arguments{
\item{tb_Grades}{A data frame return from convert_gradeList2dataframe()}

\item{disputeFolder}{A character of path to dispute folder}

\item{title}{A character of exercise title such as hw1, midterm1}
}
\value{

}
\description{
Update tb_Grades according to disputeFolder Rmd result
}
\examples{
none
}
