% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{allequalService}
\alias{allequalService}
\title{Service initiator for all.equal comparisons}
\usage{
allequalService(
  targetLabel = targetLabel,
  .transform = NULL,
  switchTargetCurrent = F
)
}
\value{

}
\description{
Service initiator for all.equal comparisons
}
\examples{
none
}
