% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{grade_by_all.equalMessages}
\alias{grade_by_all.equalMessages}
\title{Extract grades and comments from an all.equal_env grading result.}
\usage{
grade_by_all.equalMessages(ae)
}
\arguments{
\item{ae}{An all.equal_env object after assigning grades for all messages groups}
}
\value{
A list of grades and comments
}
\description{
Extract grades and comments from an all.equal_env grading result.
}
\examples{
none
}
